Spliting components and moving logic to custom hook
